lwm2m: preparing certificate verification  for bootstrap
